requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
cssselect2==0.8.0
distro==1.9.0
dj-database-url==3.1.0
Django==6.0.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
fastavro==1.12.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-core==1.2.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
MarkupSafe==3.0.3
openai==2.16.0
openapi-codec==1.3.2
orjson==3.11.5
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
PyJWT==2.9.0
pyphen==0.17.2
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.9
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==82.0.0
shellingham==1.5.4
simplejson==3.20.2
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.2
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
weasyprint==68.0
webencodings==0.5.1
whitenoise==6.11.0
xxhash==3.6.0
zopfli==0.4.0
zstandard==0.25.0
